requirements.txt
# S4Tool-Image ComfyUI Custom Node Requirements
# Production quality dependencies - all packages are mandatory

# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
Pillow>=8.0.0
numpy>=1.21.0

# Scientific computing and image processing
scikit-learn>=1.0.0
scikit-image>=0.19.0
scipy>=1.8.0
opencv-contrib-python>=4.5.0

# AI model support (for RMBG-2.0 background removal)
transformers>=4.20.0
huggingface_hub>=0.15.0
safetensors>=0.3.0